IndianLegal/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pymupdf==1.23.26
python-docx==1.1.0
openai==1.14.0
spacy==3.7.4
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
pydantic==2.6.3
streamlit==1.32.2
streamlit-option-menu==0.3.12
plotly==5.19.0 